llm_crud2_copy 1/llm_crud2_copy/requirements.txt
streamlit>=1.29.0
psycopg2-binary>=2.9.9
openai>=1.3.8
python-dotenv>=1.0.0
sqlparse>=0.4.4
pandas>=2.1.4
fastmcp>=0.2.0
pydantic>=2.5.3
groq
Faker